+++ src/Odotocodot.OneNote.Linq/Abstractions/INavigable.cs
@@ -1,0 +1,14 @@
+namespace Odotocodot.OneNote.Linq.Abstractions
+{
+    /// <summary>
+    /// Represents an object that can be opened in OneNote.
+    /// </summary>
+    /// <seealso cref="OneNoteApplication.OpenInOneNote(INavigable)"/>
+    public interface INavigable
+    {
+        /// <summary>
+        /// The ID of the OneNote hierarchy item.
+        /// </summary>
+        string Id { get; }
+    }
+}