+++ src/Odotocodot.OneNote.Linq/Root.cs
@@ -8,9 +8,11 @@ namespace Odotocodot.OneNote.Linq
 	{
 		public IReadOnlyList<Notebook> Notebooks { get; internal set; }
 
-		//public UnfiledNotes UnfiledNotes { get; internal set; }
+#nullable enable
+		public UnfiledNotes? UnfiledNotes { get; internal set; }
 
-		//public OpenSections OpenSections { get; internal set; }
+		public OpenSections? OpenSections { get; internal set; }
+#nullable restore
 
 		IEnumerator<Notebook> IEnumerable<Notebook>.GetEnumerator() => Notebooks.GetEnumerator();
 		IEnumerator IEnumerable.GetEnumerator() => ((IEnumerable)Notebooks).GetEnumerator();