+++ src/Odotocodot.OneNote.Linq/UnfiledNotes.cs
@@ -1,0 +1,11 @@
+using Odotocodot.OneNote.Linq.Abstractions;
+
+namespace Odotocodot.OneNote.Linq
+{
+    public class UnfiledNotes : INavigable
+    {
+        internal UnfiledNotes() { }
+        public Section Section { get; internal set; }
+        public string Id { get; internal set; }
+    }
+}